🚚 vue-dot: Move DatePicker mixins tests to dedicated folder (#430)
