+++ packages/vue-dot/src/patterns/DatePicker/mixins/tests/birthdate.spec.ts
@@ -3,9 +3,9 @@ import { mount, Wrapper } from '@vue/test-utils';
 
 import dayjs from 'dayjs';
 
-import { Refs } from '../../../types';
+import { Refs } from '../../../../types';
 
-import { Birthdate } from '../mixins/birthdate';
+import { Birthdate } from '../birthdate';
 
 // Tell jest to mock all timeout functions
 jest.useFakeTimers();
